COLLEGE_MANAGMENT_SYSTEM/backend/python/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
mysql-connector-python==8.3.0
pydantic==2.7.1
python-multipart==0.0.9
